Move event test to a new file
